+++ test/scrapers/tvmaze/testTvMazeShowSearch.cpp
@@ -13,7 +13,7 @@ TEST_CASE("TvMaze returns valid search results", "[tv][TvMaze][search]")
     {
         ShowSearchJob::Config config{"Simpsons", Locale::English};
         auto* searchJob = new TvMazeShowSearchJob(getTvMazeApi(), config);
-        const auto scraperResults = searchTvScraperSync(searchJob).first;
+        const auto scraperResults = test::searchTvScraperSync(searchJob).first;
 
         REQUIRE(scraperResults.length() >= 2);
         CHECK(scraperResults[0].title == "The Simpsons");
@@ -25,7 +25,7 @@ TEST_CASE("TvMaze returns valid search results", "[tv][TvMaze][search]")
     {
         ShowSearchJob::Config config{"SomethingThatDoesNotExist", Locale::English};
         auto* searchJob = new TvMazeShowSearchJob(getTvMazeApi(), config);
-        const auto p = searchTvScraperSync(searchJob, true);
+        const auto p = test::searchTvScraperSync(searchJob, true);
 
         CHECK(p.first.length() == 0);
         CHECK(p.second.error == ScraperError::Type::NoError);