+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
  - Movie renamer: It is now possible to use a movie's TMDb ID as placeholder (#1542)
  - Custom Movie Scraper: You can now load movie tags as well (#1162)
+ - A new TV scraper for fernsehserien.de was added (#1135)
 
 ### Removed